Create .travis.yml
Fixes #3 Travis CI would need to be turned ON for this repo by @malwaredllc going to https://travis-ci.com/profile and logging in with GitHub credentials.

Travis CI will run [flake8](http://flake8.pycqa.org) tests on Python 2.7 and 3.7 to find Python syntax errors and undefined names.  Python 3 will be run in __allow_failures__ mode because it is not yet a supported platform.  #32

__E901,E999,F821,F822,F823__ are the "_showstopper_" flake8 issues that can halt the runtime with a SyntaxError, NameError, etc. Most other flake8 issues are merely "style violations" -- useful for readability but they do not effect runtime safety.
* F821: undefined name `name`
* F822: undefined name `name` in `__all__`
* F823: local variable name referenced before assignment
* E901: SyntaxError or IndentationError
* E999: SyntaxError -- failed to compile a file into an Abstract Syntax Tree