Update server.py
Removed 'id' from the info table, leaving only the field in the Session object to keep track of ID.
I did this because the code already logically separates sessions from clients, and contains a method to list session IDs for a client. So it doesn't make sense for IDs to be stored permanently in the database.
They now stay in runtime (as fields in their associated sessions) and aren't stored in the DB which was causing inconsistencies.